requirements.txt
streamlit
pydeck
google-generativeai
python-dotenv
pandas
numpy